Update oc-mirror to not require handlers
The current implementation of oc-mirror in the openshift role fails on ansible-lint. This is due to the fact, that the download relied on oc_mirror_download_bundle.changed, which should be used with handlers. To avoid this issue and have a simple codebase, the condition was changed to having the variable defined.